+++ thk301/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+###################################
+#  Tae Kim
+#
+#  __init__.py 
+#  March 31,2015
+#
+#  initializer
+#
+###################################
+
